% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\docType{data}
\name{price}
\alias{price}
\title{Aggregate gram price data}
\format{
A \code{ts} object with 156 observations.
}
\source{
\url{https://www.agmarknet.gov.in/}
}
\usage{
price
}
\description{
National aggregate price of gram from Indian markets, which is a major pulse
in the country. The observations range from January, 2010 upto December,
2023.
}
\examples{
plot(price, xlab = "Year", ylab = "Aggregate price of Gram (Rs./Bag)")
}
\keyword{data}
